// frontend/src/core/api/generated/model/dayMarkView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { DayMarkViewStatus } from './dayMarkViewStatus'

export interface DayMarkView {
  entryId: string
  day: string
  status: DayMarkViewStatus
  /** @nullable */
  overtimeHours?: number | null
  /** @nullable */
  note?: string | null
  markedByName: string
  markedAt: string
}