// frontend/src/core/api/generated/model/dayMarkViewStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type DayMarkViewStatus = (typeof DayMarkViewStatus)[keyof typeof DayMarkViewStatus]

export const DayMarkViewStatus = {
  PRESENT: 'PRESENT',
  HALF_DAY: 'HALF_DAY',
  ABSENT: 'ABSENT',
  LEAVE: 'LEAVE',
} as const